perf: Allow -d to be specified twice
The -d option for UCT tests specifies the name of the device to be used
for testing. This device name is used on both the server and the client.
In some scenarios, the name of the device on the client machine might
not match the name of the device on the server machine.

Allow the user to specify -d a second time. If -d is specified only
once, we continue to get the original behavior of using the same device
name in both the client and the server. If specified a second time, then
the first instance of -d will be used as the device name on the server,
but the second instance will be used as the device name on the client.

Signed-off-by: Evan Green <evan@rivosinc.com>