UCM/BISTRO/RISCV: Fix address calculations
Loading a 64-bit address in via immediates on RISC-V is not at all
straightforward. Not only are you restricted to immediate loads of 20
and 12, but all the add operations are sign extending. This means every
time you combine two pieces of your address, if the top bit is 1, it
turns into a subtraction.

There are 3 adds in the bistro loop. To compensate for the sign
extension turning into a "borrow" on the upper bits, we need to add an
extra +1 to those upper bits. The extra adds can also carry into each
other, so they need to be done sequentially.

Signed-off-by: Evan Green <evan@rivosinc.com>