UCM/BISTRO/RISCV: Fix address calculations
Loading a 64-bit address in via immediates on RISC-V is not at all
straightforward. Not only are you restricted to immediate loads of 20
and 12, but all the add operations are sign extending. This means every
time you combine two pieces of your address, if the top bit is 1, it
turns into a subtraction.

Rather than doing all these error-prone calculations, it's simpler to
just plunk the address in the text stream, do a pc-relative load, and
jump to it. In both cases it's an indirect branch via register, takes
the same amount of space, and will be accelerated by the branch
predictor.

Signed-off-by: Evan Green <evan@rivosinc.com>